refactor get_unique_nodes_ordered_from_way
