+++ osmnx/geo_utils.py
@@ -11,6 +11,7 @@ import io
 import logging as lg
 import math
 import os
+import geopandas as gpd
 import networkx as nx
 import numpy as np
 import pandas as pd
@@ -29,7 +30,6 @@ from shapely.geometry import Polygon
 
 from .downloader import nominatim_request
 from .osm_content_handler import OSMContentHandler
-from .save_load import graph_to_gdfs
 from .utils import log
 from . import settings
 
@@ -981,3 +981,174 @@ def bbox_to_poly(north, south, east, west):
     """
 
     return Polygon([(west, south), (east, south), (east, north), (west, north)])
+
+
+def get_unique_nodes_ordered_from_way(way_edges_df):
+    """
+    Function to recover the original order of nodes from a dataframe
+    of edges associated with a single OSM way.
+
+    Parameters
+    ----------
+    way_edges_df : pandas.DataFrame()
+        Dataframe containing columns 'u' and 'v' corresponding to
+        origin/desitination nodes.
+
+    Returns
+    -------
+    unique_ordered_nodes : list
+        An ordered list of unique node IDs
+
+    NOTE: If the edges do not all connect (e.g. [(1, 2), (2,3),
+    (10, 11), (11, 12), (12, 13)]), then this method will return
+    only those nodes associated with the largest component of
+    connected edges, even if subsequent connected chunks are contain
+    more total nodes. This is done to ensure a proper topological
+    representation of nodes in the XML way records because if there
+    are unconnected components, the sorting algorithm cannot recover
+    their original order. I don't believe that we would ever encounter
+    this kind of disconnected structure of nodes within a given way,
+    but as best I could tell it is not explicitly forbidden in the
+    OSM XML design schema.
+    """
+
+    G = nx.MultiDiGraph()
+    all_nodes = list(way_edges_df['u'].values) + \
+        list(way_edges_df['v'].values)
+
+    G.add_nodes_from(all_nodes)
+    G.add_edges_from(way_edges_df[['u', 'v']].values)
+
+    # copy nodes into new graph
+    G2 = get_largest_component(G, strongly=False)
+    unique_ordered_nodes = list(nx.topological_sort(G2))
+    num_unique_nodes = len(np.unique(all_nodes))
+
+    if len(unique_ordered_nodes) < num_unique_nodes:
+        log('Recovered order for {0} of {1} nodes'.format(
+            len(unique_ordered_nodes), num_unique_nodes))
+
+    return unique_ordered_nodes
+
+
+
+def graph_to_gdfs(G, nodes=True, edges=True, node_geometry=True, fill_edge_geometry=True):
+    """
+    Convert a graph into node and/or edge GeoDataFrames
+
+    Parameters
+    ----------
+    G : networkx multidigraph
+    nodes : bool
+        if True, convert graph nodes to a GeoDataFrame and return it
+    edges : bool
+        if True, convert graph edges to a GeoDataFrame and return it
+    node_geometry : bool
+        if True, create a geometry column from node x and y data
+    fill_edge_geometry : bool
+        if True, fill in missing edge geometry fields using origin and
+        destination nodes
+
+    Returns
+    -------
+    GeoDataFrame or tuple
+        gdf_nodes or gdf_edges or both as a tuple
+    """
+
+    if not (nodes or edges):
+        raise ValueError('You must request nodes or edges, or both.')
+
+    to_return = []
+
+    if nodes:
+
+        start_time = time.time()
+
+        nodes, data = zip(*G.nodes(data=True))
+        gdf_nodes = gpd.GeoDataFrame(list(data), index=nodes)
+        if node_geometry:
+            gdf_nodes['geometry'] = gdf_nodes.apply(lambda row: Point(row['x'], row['y']), axis=1)
+            gdf_nodes.set_geometry('geometry', inplace=True)
+        gdf_nodes.crs = G.graph['crs']
+        gdf_nodes.gdf_name = '{}_nodes'.format(G.graph['name'])
+
+        to_return.append(gdf_nodes)
+        log('Created GeoDataFrame "{}" from graph in {:,.2f} seconds'.format(gdf_nodes.gdf_name, time.time()-start_time))
+
+    if edges:
+
+        start_time = time.time()
+
+        # create a list to hold our edges, then loop through each edge in the
+        # graph
+        edges = []
+        for u, v, key, data in G.edges(keys=True, data=True):
+
+            # for each edge, add key and all attributes in data dict to the
+            # edge_details
+            edge_details = {'u':u, 'v':v, 'key':key}
+            for attr_key in data:
+                edge_details[attr_key] = data[attr_key]
+
+            # if edge doesn't already have a geometry attribute, create one now
+            # if fill_edge_geometry==True
+            if 'geometry' not in data:
+                if fill_edge_geometry:
+                    point_u = Point((G.nodes[u]['x'], G.nodes[u]['y']))
+                    point_v = Point((G.nodes[v]['x'], G.nodes[v]['y']))
+                    edge_details['geometry'] = LineString([point_u, point_v])
+                else:
+                    edge_details['geometry'] = np.nan
+
+            edges.append(edge_details)
+
+        # create a GeoDataFrame from the list of edges and set the CRS
+        gdf_edges = gpd.GeoDataFrame(edges)
+        gdf_edges.crs = G.graph['crs']
+        gdf_edges.gdf_name = '{}_edges'.format(G.graph['name'])
+
+        to_return.append(gdf_edges)
+        log('Created GeoDataFrame "{}" from graph in {:,.2f} seconds'.format(gdf_edges.gdf_name, time.time()-start_time))
+
+    if len(to_return) > 1:
+        return tuple(to_return)
+    else:
+        return to_return[0]
+
+
+def gdfs_to_graph(gdf_nodes, gdf_edges):
+    """
+    Convert node and edge GeoDataFrames into a graph
+
+    Parameters
+    ----------
+    gdf_nodes : GeoDataFrame
+    gdf_edges : GeoDataFrame
+
+    Returns
+    -------
+    networkx multidigraph
+    """
+
+    G = nx.MultiDiGraph()
+    G.graph['crs'] = gdf_nodes.crs
+    G.graph['name'] = gdf_nodes.gdf_name.rstrip('_nodes')
+
+    # add the nodes and their attributes to the graph
+    G.add_nodes_from(gdf_nodes.index)
+    attributes = gdf_nodes.to_dict()
+    for attribute_name in gdf_nodes.columns:
+        # only add this attribute to nodes which have a non-null value for it
+        attribute_values = {k:v for k, v in attributes[attribute_name].items() if pd.notnull(v)}
+        nx.set_node_attributes(G, name=attribute_name, values=attribute_values)
+
+    # add the edges and attributes that are not u, v, key (as they're added
+    # separately) or null
+    for _, row in gdf_edges.iterrows():
+        attrs = {}
+        for label, value in row.iteritems():
+            if (label not in ['u', 'v', 'key']) and (isinstance(value, list) or pd.notnull(value)):
+                attrs[label] = value
+        G.add_edge(row['u'], row['v'], key=row['key'], **attrs)
+
+    return G
