+++ osmnx/utils.py
@@ -294,77 +294,3 @@ def get_logger(level=None, name=None, filename=None):
         logger.handler_set = True
 
     return logger
-
-
-def get_unique_nodes_ordered_from_way(way_edges_df):
-    """
-    Function to recover the original order of nodes from a dataframe
-    of edges associated with a single OSM way.
-
-    Parameters
-    ----------
-    way_edges_df : pandas.DataFrame()
-        Dataframe containing columns 'u' and 'v' corresponding to
-        origin/desitination nodes.
-
-    Returns
-    -------
-    unique_ordered_nodes : list
-        An ordered list of unique node IDs
-
-    NOTE: If the edges do not all connect (e.g. [(1, 2), (2,3),
-    (10, 11), (11, 12), (12, 13)]), then this method will return
-    only those nodes associated with the largest component of
-    connected edges, even if subsequent connected chunks are contain
-    more total nodes. This is done to ensure a proper topological
-    representation of nodes in the XML way records because if there
-    are unconnected components, the sorting algorithm cannot recover
-    their original order. I don't believe that we would ever encounter
-    this kind of disconnected structure of nodes within a given way,
-    but as best I could tell it is not explicitly forbidden in the
-    OSM XML design schema. I'm using a print statement right now to
-    tell the user whether or not any nodes have been dropped and
-    how many.
-    """
-
-    G = nx.MultiDiGraph()
-    all_nodes = list(way_edges_df['u'].values) + \
-        list(way_edges_df['v'].values)
-
-    G.add_nodes_from(all_nodes)
-    G.add_edges_from(way_edges_df[['u', 'v']].values)
-    wccs = nx.weakly_connected_components(G)
-    largest_wcc = max(wccs, key=len)
-    node_subset = set(largest_wcc)
-
-    # NOTE: this code (L387-403) is copied from geo_utils.py
-    # which cannot be imported here without triggering a
-    # circular import error. This should be fixed next time the
-    # code base is refactored
-
-    # copy nodes into new graph
-    G2 = G.__class__()
-    G2.add_nodes_from((n, G.nodes[n]) for n in node_subset)
-
-    # copy edges to new graph, including parallel edges
-    if G2.is_multigraph:
-        G2.add_edges_from((n, nbr, key, d)
-            for n, nbrs in G.adj.items() if n in node_subset
-            for nbr, keydict in nbrs.items() if nbr in node_subset
-            for key, d in keydict.items())
-    else:
-        G2.add_edges_from((n, nbr, d)
-            for n, nbrs in G.adj.items() if n in node_subset
-            for nbr, d in nbrs.items() if nbr in node_subset)
-
-    # update graph attribute dict, and return graph
-    G2.graph.update(G.graph)
-
-    unique_ordered_nodes = list(nx.topological_sort(G2))
-    num_unique_nodes = len(np.unique(all_nodes))
-
-    if len(unique_ordered_nodes) < num_unique_nodes:
-        log('Recovered order for {0} of {1} nodes'.format(
-            len(unique_ordered_nodes), num_unique_nodes))
-
-    return unique_ordered_nodes
